server/requirements.txt
aiohttp==3.8.5
anyio==3.7.1
async-timeout==3.0.1
attrs==23.1.0
certifi==2023.7.22
chardet==3.0.4
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==2.10
iso8601==2.0.0
metaapi-cloud-copyfactory-sdk==7.0.0
metaapi-cloud-metastats-sdk==3.2.2
metaapi-cloud-risk-management-sdk==2.0.0
metaapi-cloud-sdk==21.3.0
multidict==6.0.4
python-dotenv==1.0.0
python-engineio==3.14.2
python-socketio==4.6.0
pytz==2023.3
requests==2.31.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
typing-extensions==3.10.0.2
urllib3==1.26.5
websockets==9.1
yarl==1.9.2
